+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "triangulation"
-version = "0.1.20"
+version = "0.1.21"
 description = ""
 authors = ["laurenssam <laurenssamson@hotmail.com>"]
